refactor: split process-control ipc modules
